refactor(console): connector platform icon
