--- packages/console/src/icons/Native.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const Native = () => (
-  <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M8.4735 10.8599L8.3735 10.7799C8.33633 10.7547 8.29598 10.7345 8.2535 10.7199L8.1335 10.6666C8.02538 10.644 7.91334 10.6486 7.80745 10.68C7.70155 10.7114 7.60512 10.7687 7.52683 10.8466C7.46794 10.9114 7.42059 10.9858 7.38683 11.0666C7.33638 11.1881 7.32298 11.3217 7.34833 11.4508C7.37368 11.5799 7.43665 11.6985 7.52929 11.7919C7.62194 11.8853 7.74014 11.9491 7.869 11.9755C7.99787 12.0018 8.13164 11.9894 8.2535 11.9399C8.33316 11.9013 8.40718 11.8519 8.4735 11.7933C8.566 11.6995 8.62867 11.5805 8.65358 11.4511C8.6785 11.3218 8.66455 11.188 8.6135 11.0666C8.58026 10.9895 8.53276 10.9194 8.4735 10.8599ZM10.6668 1.33325H5.3335C4.80306 1.33325 4.29436 1.54397 3.91928 1.91904C3.54421 2.29411 3.3335 2.80282 3.3335 3.33325V12.6666C3.3335 13.197 3.54421 13.7057 3.91928 14.0808C4.29436 14.4559 4.80306 14.6666 5.3335 14.6666H10.6668C11.1973 14.6666 11.706 14.4559 12.081 14.0808C12.4561 13.7057 12.6668 13.197 12.6668 12.6666V3.33325C12.6668 2.80282 12.4561 2.29411 12.081 1.91904C11.706 1.54397 11.1973 1.33325 10.6668 1.33325ZM11.3335 12.6666C11.3335 12.8434 11.2633 13.013 11.1382 13.138C11.0132 13.263 10.8436 13.3333 10.6668 13.3333H5.3335C5.15669 13.3333 4.98712 13.263 4.86209 13.138C4.73707 13.013 4.66683 12.8434 4.66683 12.6666V3.33325C4.66683 3.15644 4.73707 2.98687 4.86209 2.86185C4.98712 2.73682 5.15669 2.66659 5.3335 2.66659H10.6668C10.8436 2.66659 11.0132 2.73682 11.1382 2.86185C11.2633 2.98687 11.3335 3.15644 11.3335 3.33325V12.6666Z"
-      fill="currentColor"
-    />
-  </svg>
-);
-
-export default Native;
