--- packages/console/src/icons/Universal.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const Universal = () => (
-  <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M6.6665 3.33341C6.53465 3.33341 6.40576 3.37251 6.29612 3.44577C6.18649 3.51902 6.10104 3.62314 6.05058 3.74496C6.00013 3.86678 5.98692 4.00082 6.01265 4.13014C6.03837 4.25946 6.10186 4.37825 6.1951 4.47149C6.28833 4.56472 6.40712 4.62821 6.53644 4.65394C6.66576 4.67966 6.79981 4.66646 6.92163 4.616C7.04344 4.56554 7.14756 4.48009 7.22082 4.37046C7.29407 4.26083 7.33317 4.13194 7.33317 4.00008C7.33317 3.82327 7.26293 3.6537 7.13791 3.52868C7.01288 3.40365 6.84331 3.33341 6.6665 3.33341ZM3.99984 3.33341C3.86798 3.33341 3.73909 3.37251 3.62946 3.44577C3.51982 3.51902 3.43438 3.62314 3.38392 3.74496C3.33346 3.86678 3.32026 4.00082 3.34598 4.13014C3.3717 4.25946 3.4352 4.37825 3.52843 4.47149C3.62167 4.56472 3.74046 4.62821 3.86978 4.65394C3.9991 4.67966 4.13314 4.66646 4.25496 4.616C4.37678 4.56554 4.4809 4.48009 4.55415 4.37046C4.6274 4.26083 4.6665 4.13194 4.6665 4.00008C4.6665 3.82327 4.59627 3.6537 4.47124 3.52868C4.34622 3.40365 4.17665 3.33341 3.99984 3.33341ZM9.33317 3.33341C9.20132 3.33341 9.07242 3.37251 8.96279 3.44577C8.85316 3.51902 8.76771 3.62314 8.71725 3.74496C8.66679 3.86678 8.65359 4.00082 8.67931 4.13014C8.70504 4.25946 8.76853 4.37825 8.86177 4.47149C8.955 4.56472 9.07379 4.62821 9.20311 4.65394C9.33243 4.67966 9.46648 4.66646 9.58829 4.616C9.71011 4.56554 9.81423 4.48009 9.88748 4.37046C9.96074 4.26083 9.99984 4.13194 9.99984 4.00008C9.99984 3.82327 9.9296 3.6537 9.80457 3.52868C9.67955 3.40365 9.50998 3.33341 9.33317 3.33341ZM13.3332 0.666748H2.6665C2.13607 0.666748 1.62736 0.877462 1.25229 1.25253C0.877218 1.62761 0.666504 2.13631 0.666504 2.66675V13.3334C0.666504 13.8638 0.877218 14.3726 1.25229 14.7476C1.62736 15.1227 2.13607 15.3334 2.6665 15.3334H13.3332C13.8636 15.3334 14.3723 15.1227 14.7474 14.7476C15.1225 14.3726 15.3332 13.8638 15.3332 13.3334V2.66675C15.3332 2.13631 15.1225 1.62761 14.7474 1.25253C14.3723 0.877462 13.8636 0.666748 13.3332 0.666748ZM13.9998 13.3334C13.9998 13.5102 13.9296 13.6798 13.8046 13.8048C13.6796 13.9298 13.51 14.0001 13.3332 14.0001H2.6665C2.48969 14.0001 2.32012 13.9298 2.1951 13.8048C2.07008 13.6798 1.99984 13.5102 1.99984 13.3334V7.33341H13.9998V13.3334ZM13.9998 6.00008H1.99984V2.66675C1.99984 2.48994 2.07008 2.32037 2.1951 2.19534C2.32012 2.07032 2.48969 2.00008 2.6665 2.00008H13.3332C13.51 2.00008 13.6796 2.07032 13.8046 2.19534C13.9296 2.32037 13.9998 2.48994 13.9998 2.66675V6.00008Z"
-      fill="currentColor"
-    />
-  </svg>
-);
-
-export default Universal;
