--- packages/console/src/icons/Web.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const Web = () => (
-  <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M12.6668 2H3.3335C2.80306 2 2.29436 2.21071 1.91928 2.58579C1.54421 2.96086 1.3335 3.46957 1.3335 4V9.33333C1.3335 9.86377 1.54421 10.3725 1.91928 10.7475C2.29436 11.1226 2.80306 11.3333 3.3335 11.3333H7.3335V12.6667H4.66683C4.49002 12.6667 4.32045 12.7369 4.19543 12.8619C4.0704 12.987 4.00016 13.1565 4.00016 13.3333C4.00016 13.5101 4.0704 13.6797 4.19543 13.8047C4.32045 13.9298 4.49002 14 4.66683 14H11.3335C11.5103 14 11.6799 13.9298 11.8049 13.8047C11.9299 13.6797 12.0002 13.5101 12.0002 13.3333C12.0002 13.1565 11.9299 12.987 11.8049 12.8619C11.6799 12.7369 11.5103 12.6667 11.3335 12.6667H8.66683V11.3333H12.6668C13.1973 11.3333 13.706 11.1226 14.081 10.7475C14.4561 10.3725 14.6668 9.86377 14.6668 9.33333V4C14.6668 3.46957 14.4561 2.96086 14.081 2.58579C13.706 2.21071 13.1973 2 12.6668 2ZM13.3335 9.33333C13.3335 9.51014 13.2633 9.67971 13.1382 9.80474C13.0132 9.92976 12.8436 10 12.6668 10H3.3335C3.15669 10 2.98712 9.92976 2.86209 9.80474C2.73707 9.67971 2.66683 9.51014 2.66683 9.33333V4C2.66683 3.82319 2.73707 3.65362 2.86209 3.5286C2.98712 3.40357 3.15669 3.33333 3.3335 3.33333H12.6668C12.8436 3.33333 13.0132 3.40357 13.1382 3.5286C13.2633 3.65362 13.3335 3.82319 13.3335 4V9.33333Z"
-      fill="currentColor"
-    />
-  </svg>
-);
-
-export default Web;
